GBA: various changes to all carts listed
EZODE has the best battery life with Revision B. This removes a selling point for the Everdrive, however, krikzz tends to offer better support so I've added that as another selling point. Add information about replacement batteries - this is especially important for flashcarts that use SRAM for saving. Add more information on custom firmwares for the SuperCard SD