sass: remove deprecated import rules
import rules will soon be deprecated in dart sass 3.0.0. We can instead use the use rule to get rid of the import rule warnings. The changes mean that we have to set default colour values in the base stylesheet then configure them in the dark/light theme stylesheets. I've set the default to be the light theme. We do not need to configure the light colour values in the light theme stylesheet, but we do need to configure the dark colour values in the dark theme stylesheet. This seems to work but feels janky, might need more looking into before merging.